rename : entities/qnaboard-coverLetter 디렉토리 구조 재정의
- api: 백엔드와의 직접적인 연결이 연관된 코드
- model: 데이터 저장소, 데이터 규칙, 상태 관리 로직
- lib: 비즈니스 로직과는 직접적인 연관이 없는 유틸성
- ui: 엔터티의 데이터를 시각적으로 표현 가능한 비즈니스 로직이 포함되지 않은 UI
