+++ client/fixutures/entities/coverLetter.ts
@@ -3,7 +3,7 @@ import {
   ICoverLetters,
   IReviseCoverLetter,
   IWriteCoverLetter,
-} from '@/src/entities/coverLetter/lib/type';
+} from '@/src/entities/coverLetter/model/type';
 
 /** 샘플 자소서 작성 목록 */
 const sampleCoverLetters: IWriteCoverLetter = {
