+++ client/src/entities/coverLetter/index.ts
@@ -1,8 +1,8 @@
 export * as coverLetterApi from './api/apis';
-export * as coverLetterMutation from './model/mutations';
-export * as coverLetterQueries from './model/queries';
-export { default as querykeys } from './lib/querykeys';
-export * as coverLetterType from './lib/type';
+export * as coverLetterMutation from './api/mutations';
+export * as coverLetterQueries from './api/queries';
+export { default as querykeys } from './api/querykeys';
+export * as coverLetterType from './model/type';
 export { default as useCoverLetterStore } from './model/useCoverLetterStore';
 export { default as CoverLetterItem } from './ui/CoverLetterItem';
 export * as coverLetterModel from './model/schema';
