+++ client/src/entities/coverLetter/model/type.ts
