+++ client/src/entities/qnaboard/api/mutaions.ts
@@ -1,7 +1,10 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import { IReviseQnaBoard, IWriteQnaBoard } from '@/src/entities/qnaboard/model/type';
+
 import { useToast } from '@/src/shared/hooks';
 import { ERROR_MESSAGES, SUCCESS_MESSAGES, TOAST_MODE } from '@/src/shared/const';
-import { IReviseQnaBoard, IWriteQnaBoard } from '@/src/entities/qnaboard/type';
+
 import board from './queryKeys';
 import { deleteQnaBoard, reviseQnaBoard, writeQnaBoard } from './apis';
