+++ client/src/entities/qnaboard/index.ts
@@ -1,9 +1,9 @@
-export * as boardApi from './apis';
-export * as boardMutation from './mutaions';
-export * as boardQueries from './queries';
-export * as boardType from './type';
+export * as boardApi from './api/apis';
+export * as boardMutation from './api/mutaions';
+export * as boardQueries from './api/queries';
+export * as boardType from './model/type';
 export * as boardSchema from './model/schema';
-export { default as queryKeys } from './queryKeys';
-export { default as useBoardStore } from './useBoardStore';
-export { default as useCorrectionStore } from './useCorrectionStore';
-export { default as useFeedStore } from './useFeedStore';
+export { default as queryKeys } from './api/queryKeys';
+export { default as useBoardStore } from './model/useBoardStore';
+export { default as useCorrectionStore } from './model/useCorrectionStore';
+export { default as useFeedStore } from './model/useFeedStore';
