+++ client/src/entities/qnaboard/model/type.ts
